Removed check_config() function
The way collectd checks a config file for errors results in it being initialized (including loading plugins, such as logfile, and writing to the collectd log). This redundant considering start-stop-daemon would fail to start which would warrant an investigation by the user.

Please note that systemd does not perform a config file check for the same reason. FreeBSD's rc scripts do not perform this action either. This would explain why the behavior was not seen on my CentOS and FreeBSD servers.